Prevent double application of force
If the player collided with two ground objects at once – which could happen since they were placed next to each other – then the OnCollisionEnter handler would be applied twice, resulting in a double application of force and the player shooting up into the sky.

Now the parent of all the tiles has the ground tag, instead of each tile having a ground tag. Since the parent is a single game object, there can only be one collision with the player at any one time, and thus only a single application of force.